move DependencyMissing from scattering to core
