+++ holopy/scattering/errors.py
@@ -81,12 +81,6 @@ class TmatrixFailure(Exception):
             reason=list(logfile)[-1]
         return("Tmatrix calculation failed. This might be because your scatterer's size or aspect ratio is too large for default parameters. \n Tmatrix error message: " + reason + "Full details are available in " + self.logfilestr)
 
-class DependencyMissing(Exception):
-    def __init__(self, dep):
-        self.dep = dep
-    def __str__(self):
-        return("External Dependency: " + self.dep + " could not be found, terminating.")
-
 class AutoTheoryFailed(Exception):
     def __init__(self, scatterer):
         self.scatterer = scatterer
